+++ country/lib/src/country.dart
@@ -1,5 +1,3 @@
-part 'country.g.dart';
-
 /// Class for storing information of the country
 class Country {
   /// Creates country
@@ -33,10 +31,12 @@ class Country {
     this.addressFormat,
     this.vatRates,
     this.nanpPrefix,
-    this.eeaMember,
-    this.euMember,
-    this.esmMember,
+    required this.eeaMember,
+    required this.euMember,
+    required this.esmMember,
     this.altCurrency,
+    required this.g7Member,
+    required this.g20Member,
     required this.isoShortNameByLocale,
     required this.subdivision,
   });
@@ -75,10 +75,12 @@ class Country {
             ? null
             : VatRates.fromJson(json['vat_rates']),
         nanpPrefix: json['nanp_prefix'],
-        eeaMember: json['eea_member'],
-        euMember: json['eu_member'],
-        esmMember: json['esm_member'],
+        eeaMember: json['eea_member'] ?? false,
+        euMember: json['eu_member'] ?? false,
+        esmMember: json['esm_member'] ?? false,
         altCurrency: json['alt_currency'],
+        g7Member: json['g7_member'] ?? false,
+        g20Member: json['g20_member'] ?? false,
         isoShortNameByLocale: json['isoShortNameByLocale'] ?? {},
         subdivision: List.from(json['subdivision'] ?? [])
             .map((e) => Subdivision.fromJson(e))
@@ -255,18 +257,24 @@ class Country {
   final String? nanpPrefix;
 
   /// If this country is a member of the European Economic Area.
-  final bool? eeaMember;
+  final bool eeaMember;
 
   /// If this country is a member of the European Union.
-  final bool? euMember;
+  final bool euMember;
 
   /// If this country is a member of the European Single Market.
-  final bool? esmMember;
+  final bool esmMember;
 
   /// A complementary currency is a currency or medium of exchange that is not
   /// necessarily a national currency, but that is thought of as supplementing
   /// or complementing national currencies.
   final String? altCurrency;
+
+  /// If this country is a member of the Group of Seven.
+  final bool g7Member;
+
+  /// If this country is a member of the Group of Twenty.
+  final bool g20Member;
 
   @Deprecated('use isoShortNameByLocale instead')
   Map<String, String> get isoShortNameByLanguage => isoShortNameByLocale;
@@ -279,7 +287,7 @@ class Country {
 
   @override
   String toString() {
-    return 'Country{alpha2: $alpha2, alpha3: $alpha3, continent: $continent, countryCode: $countryCode, currencyCode: $currencyCode, gec: $gec, geo: $geo, internationalPrefix: $internationalPrefix, ioc: $ioc, isoLongName: $isoLongName, isoShortName: $isoShortName, languagesOfficial: $languagesOfficial, languagesSpoken: $languagesSpoken, nationalDestinationCodeLengths: $nationalDestinationCodeLengths, nationalNumberLengths: $nationalNumberLengths, nationalPrefix: $nationalPrefix, nationality: $nationality, number: $number, postalCode: $postalCode, postalCodeFormat: $postalCodeFormat, region: $region, startOfWeek: $startOfWeek, subregion: $subregion, unLocode: $unLocode, unofficialNames: $unofficialNames, worldRegion: $worldRegion, addressFormat: $addressFormat, vatRates: $vatRates, nanpPrefix: $nanpPrefix, eeaMember: $eeaMember, euMember: $euMember, esmMember: $esmMember, altCurrency: $altCurrency, isoShortNameByLocale: $isoShortNameByLocale}';
+    return 'Country{alpha2: $alpha2, alpha3: $alpha3, continent: $continent, countryCode: $countryCode, currencyCode: $currencyCode, gec: $gec, geo: $geo, internationalPrefix: $internationalPrefix, ioc: $ioc, isoLongName: $isoLongName, isoShortName: $isoShortName, languagesOfficial: $languagesOfficial, languagesSpoken: $languagesSpoken, nationalDestinationCodeLengths: $nationalDestinationCodeLengths, nationalNumberLengths: $nationalNumberLengths, nationalPrefix: $nationalPrefix, nationality: $nationality, number: $number, postalCode: $postalCode, postalCodeFormat: $postalCodeFormat, region: $region, startOfWeek: $startOfWeek, subregion: $subregion, unLocode: $unLocode, unofficialNames: $unofficialNames, worldRegion: $worldRegion, addressFormat: $addressFormat, vatRates: $vatRates, nanpPrefix: $nanpPrefix, eeaMember: $eeaMember, euMember: $euMember, esmMember: $esmMember, altCurrency: $altCurrency, g7Member: $g7Member, g20Member: $g20Member, isoShortNameByLocale: $isoShortNameByLocale, subdivision: $subdivision}';
   }
 
   @override
@@ -321,7 +329,10 @@ class Country {
           euMember == other.euMember &&
           esmMember == other.esmMember &&
           altCurrency == other.altCurrency &&
-          isoShortNameByLocale == other.isoShortNameByLocale;
+          g7Member == other.g7Member &&
+          g20Member == other.g20Member &&
+          isoShortNameByLocale == other.isoShortNameByLocale &&
+          subdivision == other.subdivision;
 
   @override
   int get hashCode =>
@@ -358,7 +369,10 @@ class Country {
       euMember.hashCode ^
       esmMember.hashCode ^
       altCurrency.hashCode ^
-      isoShortNameByLocale.hashCode;
+      g7Member.hashCode ^
+      g20Member.hashCode ^
+      isoShortNameByLocale.hashCode ^
+      subdivision.hashCode;
 
   /// Emoji flags are supported on all major platforms except Windows,
   /// which displays two-letter country codes instead of emoji flag images.