+++ country_generator/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 4.0.1
+
+* added g7Member and g20Member
+* changed generated code from part of to import
+
 ## 4.0.0
 
 * support subdivision generation